// src/main/interfaces/requestIf.ts
export interface signUpRequestIf {
    userName: string;
    userId: string;
    profilePic: string;
    lobbyId: string;
    isFTUE: boolean;
    fromBack: boolean;
    dealType: string;
    gamePoolType: string;
    gameId: string;
    minPlayer: number;
    noOfPlayer: number;
    accessToken: string;
    isUseBot: boolean;
    entryFee: number;
    moneyMode: string;
    rummyType: string;
    latitude: string;
    longitude: string;
    platformCommission: number,
    gameModeId : string;
    isSplit : boolean;
    isCreateRoom : boolean;
    isReferralCode : string;
    isPrivateTableContinue?:boolean;
    isLink?:boolean;
    isCash:boolean;
    coins:number;
    cash:number;
    winCash:number;
    bonus:number;
    avatarName:string;
    botType? : string;
    botJoinTimer? : number;
    botSeatCount?:number;
}